Reworked presentation
I redid the entire introduction to give a clearer motivation for the
form of the BEMC approximation. I used the newly clarified intro to
improve the presentation of the “training” stage of BEMC. To do: figure
out and present the “power method” stage of BEMC. Figure out
appropriate constraints on the estimators.